addin: zone math moves to calendarTime — slotRanking is pure policy; smart-pick tie-break pinned by test
